Avoid saving the qualifiedNameToType for every project context
The `fromModuleToProject` function should create a project context
specific to the module that was just reviewed, and not contain data
from other modules or (in this case) from the dependencies.

This wasn't a problem until elm-review CLI v2.9.0 which attemps to
stringify the project context and save it to the file system, causing
the same data to be duplicated over and over for every module.

For instance, if analyzing a project with 1000 modules, we will end up
copying the stringified data from the dependencies 1000 times.